dom/media/test/test_bug726904.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=726904
-->

<head>
  <title>Media test: default video size</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body onload="bodyLoaded();">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=726904">Mozilla Bug 726904</a>

<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

var v1 = document.createElement("video"),
    v2 = document.createElement("video"),
    poster = "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAZAAAAGQCAAAAACl1GkQAAAAAXNSR0IArs4c6QAAAAlwSFlzAAALEwAACxMBAJqcGAAAALJJREFUeNrtwQENAAAAwqD3T20ON6AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAHg0cq4AATRk8BYAAAAASUVORK5CYII",
    resource = getPlayableVideo(gSmallTests);

function bodyLoaded(){
  // Note: For DASH, width and height would vary once the video started playing, so
  // the values would not correlate with those in manifest.js. Since this test has
  // no playing, this should not affect the result.
  is(v1.videoWidth, resource.width, "Intrinsic width should match video width");
  is(v1.videoHeight, resource.height, "Intrinsic height should match video height");
  is(v2.clientWidth, 400, "clientWidth should be 400");
  is(v2.clientHeight, 400, "clientHeight should be 400");
  SimpleTest.finish();
}

if (resource) {
  v1.poster = v2.poster = poster;

  v1.src = v2.src = "http://mochi.test:8888/tests/dom/media/test/" + resource.name;

  v1.preload = "auto";
  v2.preload = "none";

  v1.muted = v2.muted = true;

  document.body.appendChild(v1);
  document.body.appendChild(v2);
} else {
  todo(false, "No types supported");
}

</script>
</pre>
</body>
</html>
